requirements.txt
asyncio
playwright
requests
bs4
aiohttp
python-dotenv
certifi